Move IrisWebsocketService to websocket package
