+++ src/main/java/de/tum/in/www1/artemis/service/iris/websocket/IrisChatWebsocketService.java
@@ -3,7 +3,6 @@ package de.tum.in.www1.artemis.service.iris.websocket;
 import de.tum.in.www1.artemis.domain.iris.session.IrisChatSession;
 import de.tum.in.www1.artemis.domain.iris.session.IrisSession;
 import de.tum.in.www1.artemis.service.WebsocketMessagingService;
-import de.tum.in.www1.artemis.service.iris.IrisWebsocketService;
 import org.springframework.stereotype.Service;
 
 @Service
